added transaction to registration function, moved registration function to userController, removed todo and modification comments for the uesrLogin migration file to clean the code
